Use SvPVX_const() instead of SvPVX(), and const check_new().
Also change check_new()'s return type to bool, as it only returns TRUE or FALSE,
and declare it static.

Without the consting, under -DDEBUGGING, the macro for SvPVX() assigns thing,
which is const SV *, to a void *, which elicits a legitimate compiler warning.
In turn, the return value of SvPVX_const() is const char *, which means that
check_new() needs to properly declare its parameter p as const void *.