Eliminate conditional 5.11/older conditional code that is actually unnecessary.
Need to use SvRV_const() for const correctness, which isn't in ppport.h, so
provide our own fallback definition when needed.

The logic used for pre-5.11 is (locally) correct for post 5.11 too.
However, the intent of the *design* of the module is that references never get
into this part of the code. Unfortunately, in turn, the implementation of the
design isn't right, and they do. However, the fix for this is actually to fix
the design, because it too isn't right.