Move SvSTASH counting from SVt_PVCV to >= SVt_PVMG
I assume that certain stashes might not be reached using the previous
behaviour. Both objects and CVs (and possibly even others) can have a stash.
For anonymous stashes without CVs in them, I believe SvSTASH would not be
counted if an object blessed in said anonymous stash would be processed.