Attempt to count PL_compiling
Often, PL_compiling will have an op_type of OP_NULL, making it being counted as
a regular base op. However, it's memory is being allocated within the
PerlInterpreter structure or in a proper global as a full COP, not just a
BASEOP. Therefore it's possible for it to be counted wrong right now.