Add RECURSE_INTO_ALL so perl_size() can bypass refcnt
Without this perl_size() won't see into stashes that contain items that
refer to the stash until it does the areana sweep. That works but leaves
*lots* of nodes with poor names.

This is much better from a top-level perspective. It might need refining
because now addr links will appear after the item rather than before.